# server/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: PlanTogether Server
  version: 0.1.0
  description: |
    Launch MVP 서버 API. 기준 설계는 docs/account_backend_design.md다.

    인증 endpoint(/v1/auth/*)는 §6의 Idempotency-Key 규칙 밖이다. 로그인
    시점에는 idempotency 키를 묶을 user_id·device_id가 없다.
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8080
    description: 로컬 개발
paths:
...
  /v1/me:
    get:
      summary: 내 계정 조회
      operationId: getMe
      tags: [account]
      responses:
        "200":
          description: 계정 정보. ETag에 version이 있다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "423":
          $ref: "#/components/responses/AccountLocked"
    patch:
      summary: 프로필 수정
      description: |
        현재 바꿀 수 있는 것은 표시 이름뿐이다. 제어 문자와 bidi override 같은 서식
        문자를 지우고 50자로 자른다. 남는 것이 없으면 400이다.
      operationId: updateMe
      tags: [account]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: 수정됨. 재전송이면 Idempotent-Replayed 헤더가 있고 본문은 현재 상태다.
          headers:
            Idempotent-Replayed:
              $ref: "#/components/headers/IdempotentReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "409":
          $ref: "#/components/responses/Conflict"
        "423":
          $ref: "#/components/responses/AccountLocked"
    delete:
      summary: 계정 삭제 요청
      description: |
        §10. 빈 body로 보낸다. 한 트랜잭션에서 계정을 deletion_requested로 바꾸고,
        모든 세션과 기기를 폐기하며 account_deletion job을 enqueue한다.

        성공하면 access token의 기기가 즉시 폐기되므로 같은 요청을 재전송해도
        idempotency replay까지 가지 않고 인증 계층에서 401 invalid_session 또는
        423 account_locked로 차단된다.
      operationId: deleteMe
      tags: [account]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "202":
          description: 삭제 요청이 접수됐다. 본문은 deletion_requested 상태의 현재 사용자다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "409":
          $ref: "#/components/responses/Conflict"
        "423":
          $ref: "#/components/responses/AccountLocked"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/devices:
    get:
      summary: 로그인된 기기 목록
      description: |
        §5.2. 폐기되지 않은 기기만 보여준다. push token은 반환하지 않는다.
      operationId: listDevices
      tags: [account]
      responses:
        "200":
          description: 기기 목록.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevicesResponse"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "423":
          $ref: "#/components/responses/AccountLocked"
...
  /v1/devices/{id}:
    delete:
      summary: 기기 폐기
      description: |
        §5.2. 같은 사용자의 어느 기기든 폐기할 수 있다. 기기와 그 세션이 즉시
        무효가 된다. 현재 기기를 지정하면 로그아웃과 같다. 다른 사용자의 기기와
        이미 폐기된 기기는 404다. 본문을 보내지 않는다.
      operationId: revokeDevice
      tags: [account]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: 폐기됨. 재전송이면 Idempotent-Replayed 헤더가 있다.
          headers:
            Idempotent-Replayed:
              $ref: "#/components/headers/IdempotentReplayed"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "423":
          $ref: "#/components/responses/AccountLocked"
  /v1/devices/{id}/push-token:
    put:
      summary: APNs token과 알림 권한 상태 등록
      description: |
        {id}는 요청한 기기 자신이어야 한다(access token의 기기). 아니면 403이다.
        push_token과 push_environment는 함께 보내거나 함께 null이다. null이면 저장된
        token을 지운다. token은 서버에서 암호화해 저장하고 어떤 응답에도 돌려주지 않는다.
      operationId: putPushToken
      tags: [account]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenRequest"
      responses:
        "200":
          description: 등록됨.
          headers:
            Idempotent-Replayed:
              $ref: "#/components/headers/IdempotentReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "423":
          $ref: "#/components/responses/AccountLocked"
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        1~128자의 출력 가능한 ASCII. 클라이언트는 UUID를 쓴다. (사용자, 기기, 키) 단위로
        30일간 결과를 보존한다. 같은 키로 같은 요청(method, route, 본문 바이트)을 다시
        보내면 원래 결과를 돌려주고, 다른 요청이면 409 idempotency_mismatch다.
      schema: {type: string, minLength: 1, maxLength: 128}
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 현재 version을 강한 ETag로 싣는다. 예 "3". 약한 ETag와 *는 받지 않는다.
      schema: {type: string}
    DeviceID:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
  headers:
    RetryAfter:
      description: 재시도까지 기다릴 초(정수, 최소 1). 429에서는 창 종료까지 남은 초다. 남은 횟수와 한도는 알리지 않는다.
      schema: {type: integer, minimum: 1}
    IdempotentReplayed:
      description: 저장된 결과를 돌려준 재전송이면 "true". 본문은 현재 리소스 상태다.
      schema: {type: string, enum: ["true"]}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 15분 수명의 access token. 서명 외에 기기 폐기와 계정 상태를 매 요청 DB에서 확인한다.
  responses:
    Conflict:
      description: |
        version_conflict(낡은 If-Match, current_version 포함) 또는
        idempotency_mismatch(같은 키를 다른 요청에 재사용).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 인증됐지만 대상 권한이 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 없거나 이 사용자가 볼 수 없는 대상. 둘을 구분하지 않는다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidRequest:
      description: 형식 또는 필드 검증 실패.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidSession:
      description: 자격이 무효다. 원인(만료, 폐기, 위조, 재사용)을 구분하지 않는다. 클라이언트는 refresh를 시도하고, refresh도 401이면 다시 로그인한다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AccountLocked:
      description: 계정이 active가 아니다(disabled, deletion_requested, deleting, deleted).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: |
        요청 제한 초과(docs/rate_limit_design.md §6). Retry-After 이후에 재시도한다. 본문은 어느 키가
        초과했는지, 한도, 남은 횟수를 담지 않는다. **세션 상실이 아니다.** 클라이언트는 로그아웃하거나
        refresh token을 폐기하지 않는다.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unavailable:
      description: Apple 같은 외부 의존성 장애, 또는 요청 제한 장치를 확인하지 못한 경우(closed scope). 사용자 자격 문제가 아니므로 클라이언트는 로그아웃하지 않고 Retry-After 이후 재시도한다.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [invalid_request, invalid_session, forbidden, not_found, version_conflict,
                 idempotency_mismatch, sync_cursor_expired, account_locked, rate_limited, internal]
        message:
          type: string
          description: 사람이 읽는 짧은 설명. 클라이언트는 code로 분기한다.
        request_id:
          type: string
          format: uuid
        current_version:
          type: integer
          format: int64
          description: version_conflict에서만 있다.
...
    User:
      type: object
      required: [id, display_name, status, version, created_at]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        display_name: {type: string}
        status:
          type: string
          enum: [active, disabled, deletion_requested, deleting, deleted]
        version: {type: integer, format: int64}
        created_at: {type: string, format: date-time}
    MeResponse:
      type: object
      required: [request_id, user, notification_ref_key]
      additionalProperties: false
      properties:
        request_id: {type: string, format: uuid}
        user:
          $ref: "#/components/schemas/User"
        notification_ref_key:
          $ref: "#/components/schemas/NotificationRefKey"
    Device:
      type: object
      required: [id, platform, current, push_authorization, time_sensitive_setting,
                 last_seen_at, version, created_at]
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        platform: {type: string, enum: [ios]}
        current:
          type: boolean
          description: 이 요청을 보낸 기기인지.
        push_authorization:
          type: string
          enum: [unknown, authorized, provisional, denied]
        time_sensitive_setting:
          type: string
          enum: [enabled, disabled, not_supported]
        last_seen_at: {type: [string, "null"], format: date-time}
        version: {type: integer, format: int64}
        created_at: {type: string, format: date-time}
    DevicesResponse:
      type: object
      required: [request_id, devices]
      additionalProperties: false
      properties:
        request_id: {type: string, format: uuid}
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
    UpdateMeRequest:
      type: object
      required: [display_name]
      additionalProperties: false
      properties:
        display_name: {type: string}
    PushTokenRequest:
      type: object
      required: [push_token, push_environment, push_authorization, time_sensitive_setting]
      additionalProperties: false
      properties:
        push_token:
          type: [string, "null"]
          description: APNs device token의 hex 문자열(32~100바이트). null이면 저장된 token을 지운다.
        push_environment:
          type: [string, "null"]
          enum: [sandbox, production, null]
        push_authorization:
          type: string
          enum: [unknown, authorized, provisional, denied]
        time_sensitive_setting:
          type: string
          enum: [enabled, disabled, not_supported]
    DeviceResponse:
      type: object
      required: [request_id, device]
      additionalProperties: false
      properties:
        request_id: {type: string, format: uuid}
        device:
          $ref: "#/components/schemas/Device"
    NotificationRefKey:
      type: string
      description: |
        본인 전용 32바이트 키(base64url, 패딩 없음). 알림 payload의 party_ref·entity_ref를
        역조회하는 데 쓴다(notification_design.md §4.3). 공유 keychain access group에만
        저장하고 로그에 남기지 않는다.
...